requirements.txt
asgiref==3.2.10
requests
Django==3.0.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django_extensions
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
python-dotenv==0.19.1